# docs/swagger.yaml
basePath: /
definitions:
...
  models.ShiftWeek:
    properties:
      created_at:
        type: string
      department_id:
        type: integer
      end_date:
        type: string
      id:
        type: integer
      notes:
        type: string
      start_date:
        type: string
      status:
        type: string
      updated_at:
        type: string
    type: object
  responses.APIResponse:
    properties:
      data: {}
      error:
        example: Fehler bei der Verarbeitung
        type: string
      message:
        example: Operation erfolgreich
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Carsten Bröckert
  description: REST-API für die Verwaltung von Schichtplänen
  license:
    name: MIT
  title: Schichtplaner API
  version: "1.0"
paths:
...
  /api/v1/shiftweeks:
    get:
      consumes:
      - application/json
      description: Ruft alle Schichtwochen mit relevanten Beziehungen ab
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ShiftWeek'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Alle Schichtwochen abrufen
      tags:
      - shiftweeks
    post:
      consumes:
      - application/json
      description: Erstellt eine neue Schichtwoche mit Validierungen
      parameters:
      - description: Schichtwoche-Daten
        in: body
        name: shiftweek
        required: true
        schema:
          $ref: '#/definitions/models.ShiftWeek'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ShiftWeek'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Schichtwoche erstellen
      tags:
      - shiftweeks
  /api/v1/shiftweeks/{id}:
    delete:
      consumes:
      - application/json
      description: Löscht eine Schichtwoche und zugehörige Schichttage
      parameters:
      - description: Schichtwoche-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Schichtwoche löschen
      tags:
      - shiftweeks
    get:
      consumes:
      - application/json
      description: Ruft eine spezifische Schichtwoche mit allen Details ab
      parameters:
      - description: Schichtwoche-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ShiftWeek'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Einzelne Schichtwoche abrufen
      tags:
      - shiftweeks
    put:
      consumes:
      - application/json
      description: Aktualisiert eine bestehende Schichtwoche
      parameters:
      - description: Schichtwoche-ID
        in: path
        name: id
        required: true
        type: integer
      - description: Aktualisierte Schichtwoche-Daten
        in: body
        name: shiftweek
        required: true
        schema:
          $ref: '#/definitions/models.ShiftWeek'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ShiftWeek'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Schichtwoche aktualisieren
      tags:
      - shiftweeks
  /api/v1/shiftweeks/{id}/stats:
    get:
      consumes:
      - application/json
      description: Ruft statistische Daten einer Schichtwoche ab
      parameters:
      - description: Schichtwoche-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  additionalProperties: true
                  type: object
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Statistiken einer Schichtwoche abrufen
      tags:
      - shiftweeks
  /api/v1/shiftweeks/{id}/status:
    put:
      consumes:
      - application/json
      description: Aktualisiert den Status einer Schichtwoche (draft/published/archived)
      parameters:
      - description: Schichtwoche-ID
        in: path
        name: id
        required: true
        type: integer
      - description: Status-Update
        in: body
        name: status
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ShiftWeek'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.APIResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Status einer Schichtwoche aktualisieren
      tags:
      - shiftweeks
  /api/v1/shiftweeks/copy/{id}:
    post:
      consumes:
      - application/json
      description: Erstellt eine Kopie einer bestehenden Schichtwoche
      parameters:
      - description: Quell-Schichtwoche-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.ShiftWeek'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Schichtwoche kopieren
      tags:
      - shiftweeks
  /api/v1/shiftweeks/department/{id}:
    get:
      consumes:
      - application/json
      description: Ruft alle Schichtwochen einer spezifischen Abteilung ab
      parameters:
      - description: Abteilungs-ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/responses.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ShiftWeek'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.APIResponse'
      summary: Schichtwochen einer Abteilung abrufen
      tags:
      - shiftweeks
swagger: "2.0"